refactor: 각 도메인 패키지 하위에 service 폴더 생성 및 서비스 클래스 이동
