+++ apps/server/src/main/java/com/weanet/server/api/route/RouteController.java
@@ -1,5 +1,6 @@
 package com.weanet.server.api.route;
 
+import com.weanet.server.api.route.service.RouteService;
 import com.weanet.server.api.route.dto.request.RouteSaveRequest;
 import com.weanet.server.api.route.dto.response.RouteEnrichedResponse;
 import com.weanet.server.api.route.dto.response.RouteSearchResponse;
