+++ apps/server/src/main/java/com/weanet/server/api/route/service/RouteEnrichmentService.java
@@ -1,8 +1,11 @@
-package com.weanet.server.api.route;
+package com.weanet.server.api.route.service;
 
 import com.weanet.server.api.route.domain.TransportType;
 import com.weanet.server.api.route.dto.response.RouteStepResponse;
-import com.weanet.server.dto.SubwayRealtimeResponse;
+import com.weanet.server.api.route.service.SubwayService;
+import com.weanet.server.api.route.service.CongestionService;
+import com.weanet.server.api.weather.service.WeatherService;
+import com.weanet.server.api.route.dto.response.SubwayRealtimeResponse;
 import com.weanet.server.api.weather.dto.WeatherResponse;
 import com.weanet.server.api.common.util.KmaCoordinateConverter;
 import lombok.RequiredArgsConstructor;
