+++ apps/server/src/main/java/com/weanet/server/api/route/service/SubwayService.java
@@ -1,4 +1,4 @@
-package com.weanet.server.api.route;
+package com.weanet.server.api.route.service;
 
 import com.weanet.server.dto.SubwayRealtimeResponse;
 import lombok.RequiredArgsConstructor;
